Split MCP tool implementations by domain
Refactor mcp_server.py by extracting tool implementations into separate domain modules (seating, orders, delivery, check_in, public) with shared utilities. ChampagneFestivalMcpBackend now serves as a thin assembler handling auth and delegation. No changes to registered tool surface or behavior—purely organizational.
